+++ module/macros.js
@@ -1,5 +1,7 @@
 import {expertDivination} from './macros/classes/ExpertDivination.js';
+import {rayOfEnfeeblement} from './macros/spells/ray-of-enfeeblement.js';
 
 export let macros = {
-    'expertDivination': expertDivination
+    'expertDivination': expertDivination,
+    'rayOfEnfeeblement': rayOfEnfeeblement
 }